+++ files/RELEASES.md
@@ -100,3 +100,13 @@ Based on:
 - [csharp v4.0.0] files
 ### Releases
 - [NuGet v4.0.0] https://www.nuget.org/packages/WingspanFiles/4.0.0 - files
+
+## 2024-02-12 01:23:20
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.177.1 (2.252.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.1.0] files
+### Releases
+- [NuGet v4.1.0] https://www.nuget.org/packages/WingspanFiles/4.1.0 - files